refactor error functions after feedback
