+++ src/js/auth/singup/signupValidation.mjs
@@ -1,12 +1,5 @@
 import { authenticateUser } from "../../commonFunctions/api/auth/signup.mjs";
-import {
-  checkUsername,
-  checkEmail,
-  checkPasswordLength,
-  checkPasswordsMatch,
-  removeError,
-  displayError,
-} from "../../commonFunctions/errorHandling.mjs";
+import * as validation from "../../commonFunctions/errorHandling/index.mjs";
 import { handleSubmission } from "./submission.mjs";
 
 //display errors or hide error messages on key up
@@ -23,25 +16,30 @@ const PASSWORD_REPEAT_FIELD = document.querySelector("#password-repeat");
 const PASSWORD_MATCH_ERROR = document.querySelector("#pass-match-error");
 
 USERNAME_FIELD.addEventListener("keyup", () => {
-  if (checkUsername(USERNAME_FIELD.value)) {
-    removeError(USERNAME_FIELD, USERNAME_ERROR);
+  if (validation.checkUsername(USERNAME_FIELD.value)) {
+    validation.removeError(USERNAME_FIELD, USERNAME_ERROR);
   }
 });
 
 EMAIL_FIELD.addEventListener("keyup", () => {
-  if (checkEmail(EMAIL_FIELD.value)) {
-    removeError(EMAIL_FIELD, EMAIL_ERROR);
+  if (validation.checkEmail(EMAIL_FIELD.value)) {
+    validation.removeError(EMAIL_FIELD, EMAIL_ERROR);
   }
 });
 
 PASSWORD_FIELD.addEventListener("keyup", () => {
-  if (checkPasswordLength(PASSWORD_FIELD.value)) {
-    removeError(PASSWORD_FIELD, PASSWORD_ERROR);
+  if (validation.checkPasswordLength(PASSWORD_FIELD.value)) {
+    validation.removeError(PASSWORD_FIELD, PASSWORD_ERROR);
   }
 });
 
 PASSWORD_REPEAT_FIELD.addEventListener("keyup", () => {
-  if (checkPasswordsMatch(PASSWORD_FIELD.value, PASSWORD_REPEAT_FIELD.value)) {
+  if (
+    validation.checkPasswordsMatch(
+      PASSWORD_FIELD.value,
+      PASSWORD_REPEAT_FIELD.value
+    )
+  ) {
     PASSWORD_REPEAT_FIELD.classList.remove("border-danger");
     PASSWORD_MATCH_ERROR.classList.add("hidden");
   }
@@ -56,34 +54,40 @@ export function validateSignup(event) {
   event.preventDefault();
 
   if (
-    checkUsername(USERNAME_FIELD.value) &&
-    checkEmail(EMAIL_FIELD.value) &&
-    checkPasswordLength(PASSWORD_FIELD.value) &&
-    checkPasswordsMatch(PASSWORD_FIELD.value, PASSWORD_REPEAT_FIELD.value)
+    validation.checkUsername(USERNAME_FIELD.value) &&
+    validation.checkEmail(EMAIL_FIELD.value) &&
+    validation.checkPasswordLength(PASSWORD_FIELD.value) &&
+    validation.checkPasswordsMatch(
+      PASSWORD_FIELD.value,
+      PASSWORD_REPEAT_FIELD.value
+    )
   ) {
     authenticateUser();
     handleSubmission();
   } else {
-    if (checkUsername(USERNAME_FIELD.value)) {
-      removeError(USERNAME_FIELD, USERNAME_ERROR);
+    if (validation.checkUsername(USERNAME_FIELD.value)) {
+      validation.removeError(USERNAME_FIELD, USERNAME_ERROR);
     } else {
-      displayError(USERNAME_FIELD, USERNAME_ERROR);
+      validation.displayError(USERNAME_FIELD, USERNAME_ERROR);
     }
 
-    if (checkEmail(EMAIL_FIELD.value)) {
-      removeError(EMAIL_FIELD, EMAIL_ERROR);
+    if (validation.checkEmail(EMAIL_FIELD.value)) {
+      validation.removeError(EMAIL_FIELD, EMAIL_ERROR);
     } else {
-      displayError(EMAIL_FIELD, EMAIL_ERROR);
+      validation.displayError(EMAIL_FIELD, EMAIL_ERROR);
     }
 
-    if (checkPasswordLength(PASSWORD_FIELD.value)) {
-      removeError(PASSWORD_FIELD, PASSWORD_ERROR);
+    if (validation.checkPasswordLength(PASSWORD_FIELD.value)) {
+      validation.removeError(PASSWORD_FIELD, PASSWORD_ERROR);
     } else {
-      displayError(PASSWORD_FIELD, PASSWORD_ERROR);
+      validation.displayError(PASSWORD_FIELD, PASSWORD_ERROR);
     }
 
     if (
-      checkPasswordsMatch(PASSWORD_FIELD.value, PASSWORD_REPEAT_FIELD.value)
+      validation.checkPasswordsMatch(
+        PASSWORD_FIELD.value,
+        PASSWORD_REPEAT_FIELD.value
+      )
     ) {
       PASSWORD_REPEAT_FIELD.classList.remove("border-danger");
       PASSWORD_MATCH_ERROR.classList.add("hidden");
