+++ src/js/commonFunctions/errorHandling/remove.mjs
@@ -1,0 +1,10 @@
+/**
+ * Updates field styles to convey
+ * field is free of errors
+ * @param {HTMLInputElement} inputFieldOne
+ * @param {HTMLParagraphElement} inputFieldTwo
+ */
+export function removeError(inputFieldOne, inputFieldTwo) {
+  inputFieldOne.classList.remove("border-danger");
+  inputFieldTwo.classList.remove("text-danger");
+}
